+++ DummyFishValues.md
@@ -1,0 +1,6 @@
+
+INSERT INTO `cryptoquarium_db`.`fishes` (`species`, `codeSpecies`, `color1r`, `color1b`, `color1g`, `color2r`, `color2b`, `color2g`, `degree`, `percent`, `randomizeVar`, `quantityAvailable`, `price`, `createdAt`, `updatedAt`) VALUES ('Fish', 'Fish', '255', '94', '0', '2', '83', '235', '145', '72', '{color: [color2r, color2b, color2g], degree: true, percent: true}', '1000', '1', '2018-1-7', '2018-1-7');
+INSERT INTO `cryptoquarium_db`.`fishes` (`species`, `codeSpecies`, `color1r`, `color1b`, `color1g`, `color2r`, `color2b`, `color2g`, `degree`, `percent`, `randomizeVar`, `quantityAvailable`, `price`, `createdAt`, `updatedAt`) VALUES ('Puffer Fish', 'PufferFish', '0', '0', '0', '242', '237', '136', '180', '50', '{color: [color2r, color2b, color2g], degree: false, percent: false}', '1000', '3', '2018-1-7', '2018-1-7');
+
+INSERT INTO `cryptoquarium_db`.`fishes` (`species`, `codeSpecies`, `color1r`, `color1b`, `color1g`, `color2r`, `color2b`, `color2g`, `degree`, `percent`, `randomizeVar`, `quantityAvailable`, `price`, `createdAt`, `updatedAt`) VALUES ('Fish', 'Fish', '255', '94', '0', '2', '83', '235', '145', '72', '{color: [color2r, color2b, color2g], degree: true, percent: true}', '1000', '5', '2018-1-7', '2018-1-7');
+INSERT INTO `cryptoquarium_db`.`fishes` (`species`, `codeSpecies`, `color1r`, `color1b`, `color1g`, `color2r`, `color2b`, `color2g`, `degree`, `percent`, `randomizeVar`, `quantityAvailable`, `price`, `createdAt`, `updatedAt`) VALUES ('Puffer Fish', 'PufferFish', '0', '0', '0', '242', '237', '136', '180', '50', '{color: [color2r, color2b, color2g], degree: false, percent: false}', '1000', '7', '2018-1-7', '2018-1-7');
