+++ src/components/AppContainer.js
@@ -16,30 +16,48 @@ class AppContainer extends Component {
         currentPage: "Home", 
         currentBalance: 0,         
         thisUserCred: [],     
-        subTotal: 0,  
-        // cartArray: []       
+        subTotal: 0,
+        cartArray: [],
+        buyFishArray: []
     };
+
+    //ORIGINAL IN BUY COMPONENT
+    clickItem = (id) => {
+        // We are looking for the index of the target fish... so... find index of all fish where the fish is filtered to the fish with the target ID
+        // This will prevent issues when we concatenate below.  
+        let fishIndex = this.state.buyFishArray.findIndex((fish) => fish===this.state.buyFishArray.filter(fish => fish.id===id)[0]);
+        
+        this.setState({            
+            cartArray: this.state.cartArray.concat([this.state.buyFishArray[fishIndex]])
+        }, (state) => {
+            this.updateSubtotalState(this.state.subTotal + this.state.buyFishArray[fishIndex].price); 
+        });             
+    }    
+    
+    //ORIGINAL IN BUY COMPONENT
+    updateBuyFishArrayState = () => {
+        axios.get('/api/allFishTemplates')
+        .then((allfish) => {    
+            // console.log(allfish);
+            allfish.data.forEach((fish) => {
+                this.setState({buyFishArray: this.state.buyFishArray.concat([fish])})
+            })
+        })
+        .catch((err)=> {
+            console.log(err)
+        })
+    } 
     
     //FUNCTION TO HANDLE THE SUBTOTAL MATH
     updateSubtotalState = (value) => {
-        console.log("am i getting here?")
-        this.setState({subTotal: value});
-        // Loop over the cart array to find the price of each item in there
-        // this.state.cartArray.forEach((item) => {
-        //     console.log("this is item price")
-        //     console.log(item.price);
-
-        //     //Add the price (you will need to burrow into each fish object to grab the price)
-        //     const finalSubtotal = this.state.subTotal + item.price
-
-        //     //Update the state of the subtotal
-        //     this.setState({subTotal: value})         
-        // });          
-    }    
+        console.log("am i getting here? NOW")
+        this.setState({subTotal: value});               
+    }      
 
     //FUNCTION TO GET LOGGEDIN USER CREDENTIALS FROM THE DATABASE
     componentDidMount() {
-        axios.get('/api/getUserAuthenticated')
+        // axios.get('/api/getUserAuthenticated')
+        axios.get('/api/getAuthenticatedUser')
         .then((userCredentials) => {
             console.log(`So... we theoretically have user creds?`);
             console.log(userCredentials.data);
@@ -54,8 +72,10 @@ class AppContainer extends Component {
             console.log(err)
         })
     }
-
+    
     //FUNCTION FOR HANDLING ACCOUNT MATH ON CHECKOUT CLICK
+    //THIS FUNCTION NEEDS THE CARTARRAY
+        //I WANT TO CLEAR THE ARRAY AFTER ACCEPTED PURCHASE SO THERE IS AN EMPTY CART FOR THE NEXT PURCHASE
     updateBalanceAfterCheckout = () => {
         console.log("Am I clicking the checkout button")
        //I NEED TO PASS UP THE CART ARRAY TO EMPTY IT HERE--BUT IT IS BEING USED HEAVILY TWO COMPONENTS DOWN >:(
@@ -73,7 +93,7 @@ class AppContainer extends Component {
     //This function sets the state for the current page
     handlePageChange = page => {
         this.setState({currentPage: page});
-    };
+    };    
 
     //Series of if/else statements that will render the corresponding component
     //for the page that is selected as current page
@@ -92,10 +112,15 @@ class AppContainer extends Component {
             return <FishMarket 
                         checkoutChangeBalance = {this.updateBalanceAfterCheckout}
                         subTotal = {this.state.subTotal}
-                        updateSubtotalState = {this.updateSubtotalState}
-                        // cartArray = {this.state.cartArray}
+                        // updateSubtotalState = {this.updateSubtotalState}
+                        //I AM PASSING CLICKITEM/FISHARRAY/CARTARRAY/UPDATESUBTOTAL TO FISH MARKET
+                        cartArray = {this.state.cartArray}
+                        buyFishArray = {this.state.buyFishArray}
+                        clickItem = {this.clickItem}
+                        updateBuyFishArrayState = {this.updateBuyFishArrayState}    
+                        updateSubtotalState = {this.updateSubtotalState}                    
                     />
-        }
+            }
     }
 
     //We are giving the Navbar the current page property and passing the handlePageChange function
